Introduced cli parameter <gyro_offset_yaw> It allows to manually compensate gyro drift over time. rebased squashed.

+++ src/main/drivers/accgyro/accgyro.h
@@ -63,8 +63,8 @@ typedef struct gyroDev_s {
     extiCallbackRec_t exti;
     busDevice_t bus;
     float scale;                                            // scalefactor
-    int32_t gyroZero[XYZ_AXIS_COUNT];
-    int32_t gyroADC[XYZ_AXIS_COUNT];                        // gyro data after calibration and alignment
+    float gyroZero[XYZ_AXIS_COUNT];
+    float gyroADC[XYZ_AXIS_COUNT];                        // gyro data after calibration and alignment
     int32_t gyroADCRawPrevious[XYZ_AXIS_COUNT];
     int16_t gyroADCRaw[XYZ_AXIS_COUNT];
     int16_t temperature;
